refactor : external 패키지 역할에 따라 분리(s3, oauth, redis)
