+++ nmnb-application/src/main/kotlin/nmnb/application/global/auth/service/AuthServiceImpl.kt
@@ -2,7 +2,7 @@ package nmnb.application.global.auth.service
 
 import nmnb.application.global.auth.service.dto.response.AuthTokenResponse
 import nmnb.application.global.auth.service.dto.response.AuthUserResponse
-import nmnb.application.global.infrastructure.external.OAuthClientComposite
+import nmnb.application.global.infrastructure.external.oauth.OAuthClientComposite
 import nmnb.application.global.infrastructure.security.JwtTokenProvider
 import nmnb.common.properties.S3Properties
 import nmnb.domain.auth.SocialType
