+++ nmnb-application/src/test/kotlin/nmnb/application/global/auth/service/AuthServiceImplTest.kt
@@ -3,8 +3,8 @@ package nmnb.application.global.auth.service
 import nmnb.application.IntegrationTestSupport
 import nmnb.application.global.auth.service.dto.KakaoAccount
 import nmnb.application.global.auth.service.dto.KakaoProfile
-import nmnb.application.global.infrastructure.external.KakaoOAuthClient
-import nmnb.application.global.infrastructure.external.OAuthClientComposite
+import nmnb.application.global.infrastructure.external.oauth.KakaoOAuthClient
+import nmnb.application.global.infrastructure.external.oauth.OAuthClientComposite
 import nmnb.common.domain.SignUpStatus
 import nmnb.domain.auth.SocialType
 import nmnb.domain.user.User
