+++ nmnb-webflux/src/main/kotlin/nmnb/webflux/domain/post/service/PostUploadServiceImpl.kt
@@ -6,8 +6,8 @@ import nmnb.r2dbc.post.R2dbcPost
 import nmnb.r2dbc.post.R2dbcPostRepository
 import nmnb.r2dbc.user.R2dbcUser
 import nmnb.webflux.domain.post.service.dto.request.PostInfoServiceRequest
-import nmnb.webflux.global.infrastructure.external.S3Service
-import nmnb.webflux.global.infrastructure.external.ThumbnailJobProducer
+import nmnb.webflux.global.infrastructure.external.redis.ThumbnailJobProducer
+import nmnb.webflux.global.infrastructure.external.s3.S3Service
 import org.springframework.http.codec.multipart.FilePart
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
